cache requests and rename utils.ts
