+++ main.ts
@@ -1,6 +1,6 @@
 import { Plugin } from 'obsidian';
 import { parseRepoURL } from 'src/gitServerInterface.ts';
-import { getPluginsDir } from 'src/utils.ts';
+import { getPluginsDir } from 'src/pluginActions.ts';
 import GaloreSettingTab from 'src/settingsPage.ts';
 
 export default class Galore extends Plugin {
